requirements.txt
gradio
openai-whisper
numpy
sounddevice
soundfile
keyboard
